apps/ai/requirements.txt
fastapi==0.115.0
openai>=1.30.0
uvicorn[standard]==0.30.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-chroma>=1.0.0
chromadb>=0.5.0
pypdf==4.2.0
python-docx==1.1.2
python-multipart==0.0.9
httpx>=0.27.0
pydantic-settings==2.2.1
